+++ SampleAndHoldSim/src/Plugin/Plugin.cs
@@ -12,11 +12,12 @@ namespace SampleAndHoldSim
     [BepInDependency(Compatibility.Multfunction_mod_Patch.GUID, BepInDependency.DependencyFlags.SoftDependency)]
     [BepInDependency(Compatibility.PlanetMiner.GUID, BepInDependency.DependencyFlags.SoftDependency)]
     [BepInDependency(Compatibility.DSP_Battle_Patch.GUID, BepInDependency.DependencyFlags.SoftDependency)]
+    [BepInDependency(Compatibility.CheatEnabler_Patch.GUID, BepInDependency.DependencyFlags.SoftDependency)]
     public class Plugin : BaseUnityPlugin
     {
         public const string GUID = "starfi5h.plugin.SampleAndHoldSim";
         public const string NAME = "SampleAndHoldSim";
-        public const string VERSION = "0.5.5";
+        public const string VERSION = "0.5.6";
         public static Plugin instance;
         Harmony harmony;
 
@@ -73,6 +74,7 @@ namespace SampleAndHoldSim
 
         public void OnDestroy()
         {
+            Compatibility.OnDestory();
             harmony.UnpatchSelf();
             UIcontrol.OnDestory();
             UIstation.OnDestory();